Refactor server to get get external host names from environment variables
