+++ package.json
@@ -37,6 +37,7 @@
     "cookie-parser": "1.4.5",
     "dayjs": "1.10.4",
     "deep-freeze": "0.0.1",
+    "dotenv": "8.2.0",
     "express": "4.17.1",
     "express-http-proxy": "1.6.2",
     "final-form": "4.20.2",
@@ -128,7 +129,6 @@
     "clean-webpack-plugin": "4.0.0-alpha.0",
     "css-loader": "5.2.1",
     "cypress": "7.6.0",
-    "dotenv": "8.2.0",
     "dotenv-webpack": "7.0.2",
     "enzyme": "3.11.0",
     "eslint": "7.26.0",
